+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - support for headers in streamable MCP servers
 - added search to input history
+- added Langfuse OpenTelemetry instrumentation for Agent and Aider
 
 ## [0.23.0]